requirements.txt
python-dotenv~=0.21.0
pyzabbix~=1.2.1